requirements.txt
fastapi
uvicorn
pytest
httpx
email-validator
google-genai
python-dotenv
faiss-cpu
numpy